Set facility name in CCM chart
The Packet CCM allows specifying the Packet facility in two ways:
Using an env var and using the Packet metadata service.

When deploying a Packet shoot with a seed which runs outside of
Packet, the CCM which runs in the seed is crashlooping because it
can't communicate with the Packet metadata service to figure out the
facility automatically.

Since the Spec.Region key of the Shoot resource seems to be required,
looks like we can always specify the facility explicitly. Doing so
allows using non-Packet seeds and shouldn't have any effect for
Packet seeds as far as I can tell.