pkg/webhook/controlplaneexposure: add --advertise-address to API server (#115)
This commit adds EnsureKubeAPIServerDeployment implementation to
controlplane exposure webhook, which ensures, that kube-apiserver has
--advertise-address flag set to kube-apiserver Service LoadBalancer IP
address, which should be a EIP assigned by MetalLB, for clusters which
use this extension as a Seed cluster.

By default in Gardener, the IP address advertised by kube-apiserver
is an IP address of kube-apiserver Pod running on Seed cluster. This
means in order to make Shoot pods communicate with API, they must be
able to communicate with Seed Pod CIDR, which is not the case on Packet.

To make Pods connectivity to API server work on Shoot clusters, we
configure kube-apiserver to advertise EIP address to Pods. However, we
only do this for clusters using this extension as a Seed, not for all
clusters using this extension, hence the webhook is placed in
pkg/webhook/controlplaneexposure and not pkg/webhook/controlplane
package.

Closes #110

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>